backend/requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
requests
python-dotenv
python-dotenv>=1.0.0